requirements.txt
streamlit>=1.28.0,<2.0.0
dspy-ai>=2.4.0,<4.0.0
anthropic>=0.7.0,<1.0.0
PyPDF2>=3.0.0,<4.0.0
PyMuPDF>=1.23.0,<2.0.0
reportlab>=4.0.0,<4.1.0
python-docx>=0.8.11,<1.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0
Pillow>=10.0.0,<11.0.0
beautifulsoup4>=4.12.0,<5.0.0
chromadb>=0.4.0,<0.5.0
openai==1.99.5
